+++ crates/binding/src/js_plugin.rs
@@ -6,7 +6,7 @@ use crate::tsfn::{LoadResult, ReadMessage, TsFnHooks, WriteRequest};
 pub struct JsPlugin {
     pub hooks: TsFnHooks,
 }
-use mako::ast_2::file::Content;
+use mako::ast::file::Content;
 use mako::compiler::Context;
 use mako::plugin::{Plugin, PluginGenerateEndParams, PluginLoadParam};
 use mako_core::anyhow::{anyhow, Result};
