+++ crates/mako/src/lib.rs
@@ -4,7 +4,7 @@
 #![feature(result_option_inspect)]
 
 mod analyze_deps;
-pub mod ast_2;
+pub mod ast;
 mod build;
 mod chunk;
 mod chunk_graph;
