+++ crates/mako/src/test_helper.rs
@@ -4,8 +4,8 @@ use std::sync::Arc;
 
 use mako_core::tracing_subscriber::{fmt, EnvFilter};
 
-use crate::ast_2::file::File;
-use crate::ast_2::js_ast::JsAst;
+use crate::ast::file::File;
+use crate::ast::js_ast::JsAst;
 use crate::compiler::{self, Compiler, Context};
 use crate::config::{Config, Mode};
 use crate::module::{Module, ModuleId, ModuleInfo};
@@ -48,7 +48,7 @@ pub fn create_mock_module(path: PathBuf, code: &str) -> Module {
 
     let context = Arc::new(Context::default());
     let mut file = File::new(path.to_string_lossy().to_string(), context.clone());
-    file.set_content(crate::ast_2::file::Content::Js(code.to_string()));
+    file.set_content(crate::ast::file::Content::Js(code.to_string()));
     let ast = JsAst::new(&file, context.clone()).unwrap();
     let module_id = ModuleId::from_path(path.clone());
     let info = ModuleInfo {
